CIRC-337 Add tests for notices rescheduling
